backend/pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.11.1",
    "bcrypt>=5.0.0",
    "dashscope>=1.25.17",
    "ddgs>=9.13.1",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-dev-utils[standard]>=1.4.6",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-qwq>=0.3.4",
    "langfuse>=4.1.0",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "mem0ai>=1.0.11",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.38",
    "structlog>=25.5.0",
]

[tool.pytest.ini_options]
# 添加项目根目录到 Python 路径，解决测试导入问题
pythonpath = [
    ".",
]
# 配置 pytest-asyncio 使用 auto 模式，自动识别 async 测试函数
asyncio_mode = "auto"
